+++ react/features/local-recording/components/LocalRecordingInfoDialog.js
@@ -11,7 +11,7 @@ import {
     getLocalParticipant
 } from '../../base/participants';
 import { connect } from '../../base/redux';
-import LocalRecordingDialog from '../../riff-platform/components/Local-Recorder/LocalRecordingDialog';
+import LocalRecordingDialog from '../../riff-platform/components/LocalRecorder/LocalRecordingDialog';
 import { statsUpdate } from '../actions';
 import { recordingController } from '../controller';
