+++ react/features/riff-platform/components/LocalRecorder/LocalRecordingDialog.js
